+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-## v0.20.1
+## v0.20.1-WIP
 
 - Removed the blank current time entry from the logs chart as it was causing confusion when used with custom time ranges.
+
+- Removed the default Go API tests timeout and added a new`ApiScenario.Timeout` option ([#3930](https://github.com/pocketbase/pocketbase/issues/3930)).
+  A negative or zero value means no tests timeout.
+  If a single API test that takes more than 3s to complete it will have a log message visible when the test fails or when [`go test -v`] flag is used.
 
 
 ## v0.20.0